Jobs running green after moving job to to subdir under deploy
